+++ ksmartteam01/src/main/resources/templates/fragments/header.html
@@ -52,11 +52,11 @@
                                     <li><a th:href="@{/goods/goodsShippingInfoInsert}">#배송정보등록</a></li>
                                     <li><a th:href="@{/goods/goodsShippingInfoModify}">#배송정보수정</a></li>
                                     <li><a href="">#배송정보삭제</a></li>
-                                    <li><a th:href="@{/goods/goodsPaymentListSearch}">#결제내역검색</a></li>
+                                    <li><a th:href="@{/goods/goodsCategoryList}">#상품카테고리목록</a></li>
                                 </ul>
                             </li>
-                            <li><a th:href="@{/goods/goodsDetail}">#상품상세</a></li>
-                            <li><a th:href="@{/goods/goodsCartList}">#장바구니</a></li>
+                            <li><a th:href="@{/user/goods/goodsDetail}">#상품상세</a></li>
+                            <li><a th:href="@{/user/goods/goodsCartList}">#장바구니</a></li>
                             <li><a href="">#상품결제</a></li>
                             
                            <li><a href="#">#계정관련</a>
