+++ ksmartteam01/src/main/resources/templates/user/goods/goodsCartList.html
@@ -62,7 +62,7 @@
                             <tbody>
                                 <tr>
                                     <td class="shoping__cart__item">
-                                        <img src="/img/cart/petCage" alt="">
+                                        <img src="/img/cart/petCage.jpg" alt="">
                                         <h5>Vegetable’s Package</h5>
                                     </td>
                                     <td class="shoping__cart__price">
@@ -84,7 +84,7 @@
                                 </tr>
                                 <tr>
                                     <td class="shoping__cart__item">
-                                        <img src="/img/cart/PetClothes" alt="">
+                                        <img src="/img/cart/PetClothes.jpg" alt="">
                                         <h5>Fresh Garden Vegetable</h5>
                                     </td>
                                     <td class="shoping__cart__price">
@@ -106,7 +106,7 @@
                                 </tr>
                                 <tr>
                                     <td class="shoping__cart__item">
-                                        <img src="/img/cart/petToy2" alt="">
+                                        <img src="/img/cart/petToy2.jpg" alt="">
                                         <h5>Organic Bananas</h5>
                                     </td>
                                     <td class="shoping__cart__price">
@@ -134,7 +134,7 @@
             <div class="row">
                 <div class="col-lg-12">
                     <div class="shoping__cart__btns">
-                        <a th:href="@{/goods/goodsListUser}" class="primary-btn cart-btn">쇼핑계속하기</a>
+                        <a th:href="@{/user/goods/goodsListUser}" class="primary-btn cart-btn">쇼핑계속하기</a>
                         <a href="#" class="primary-btn cart-btn cart-btn-right"><span class="icon_loading"></span>
                         	    장바구니 업데이트</a>
                     </div>
