+++ ksmartteam01/src/main/resources/templates/user/goods/goodsCategoryList2.html
@@ -1,0 +1,164 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+	  xmlns:layout="http://ultraq.net.nz/thymeleaf/layout"
+	  layout:decorate="~{layout/default}">
+	<!-- 사용자 정의 타이틀 설정 -->
+	<th:block layout:fragment="customTitle">
+		<title>상품카테고리목록</title>
+	</th:block>
+	
+	<th:block layout:fragment="customTopJs">
+		
+	</th:block>
+	
+	<th:block layout:fragment="customCss">
+		<style>
+			 InsertTable{
+			 	 height: 100%;
+				 text-align: center;
+				 display: flex;
+			 }
+			 table{
+				 display:table-cell;;
+				 table-layout:fixed;
+			     border: 1px solid #444444;
+			 }
+			 th, td {
+			   border: 1px solid #444444;
+			 }
+
+			 .buttonMoDel{
+				 width: 50px;
+				 height: 25px;
+			 }
+			/*
+			{
+			height: 100%;
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            justify-content: center;
+			}*/
+    	</style>
+    </th:block>
+    
+	<th:block layout:fragment="customContents">
+	 <!-- 상품 카테고리목록 시작 -->
+    <section class="shoping-cart spad">
+	    <div class="InsertTable d-flex justify-content-center">
+	     <table class="table text-center">
+		   <thead class="thead-dark">
+		    <tr>
+		      <th scope="col" class="align-middle">대분류<br>코드</th>
+		      <th scope="col" class="align-middle">대분류<br>이름</th>
+		      <th scope="col" class="align-middle">대분류<br>등록</th>
+		      <th scope="col" class="align-middle">대분류<br>수정<br>삭제</th>
+			  <th scope="col" class="align-middle">대분류<br>시각</th>
+		      <th scope="col" class="align-middle">중분류<br>코드</th>
+	          <th scope="col" class="align-middle">중분류<br>이름</th>
+			  <th scope="col" class="align-middle">중분류<br>등록</th>
+			  <th scope="col" class="align-middle">중분류<br>수정<br>삭제</th>
+			  <th scope="col" class="align-middle">중분류<br>시각</th>
+	   		</tr>
+	  	  </thead>
+		  <tbody>
+		    <tr>
+		      <td class="align-middle">1</td>
+		      <td class="align-middle">Mark</td>
+		      <td class="align-middle">
+	          </td>
+		      <td class="align-middle">
+		      	  <!-- 
+		      	  <div class="d-flex flex-row-reverse">
+			      	<button class="buttonMoDel" type="button">수정</button>
+		          	<button class="buttonMoDel" type="button">삭제</button>
+			      </div>
+			       -->
+		      	  <button class="buttonMoDel" type="button">수정</button>
+		          <button class="buttonMoDel" type="button">삭제</button>
+		      </td>
+		      <td class="align-middle">
+	          </td>
+	          <td class="align-middle">
+	          $69.00
+	          </td>
+	          <td class="align-middle">
+	          </td>
+	          <td class="align-middle">
+	          </td>
+	          <td class="align-middle">                  
+		          	<button class="buttonMoDel" type="button">수정</button>
+		          	<button class="buttonMoDel" type="button">삭제</button>
+	          </td>
+	          <td class="align-middle">
+	          </td>                     
+		    </tr>
+		    <tr>
+		      <td class="align-middle">2</td>
+		      <td class="align-middle">Jacob</td>
+		      <td class="align-middle">Thornton</td>
+		      <td class="align-middle">               
+		          	<button class="buttonMoDel" type="button">수정</button>
+		          	<button class="buttonMoDel" type="button">삭제</button>
+		      </td>
+		      <td class="align-middle">
+	          
+	          </td>
+	          <td class="align-middle">
+	          
+	          </td>
+	          <td class="align-middle">
+	          </td>
+	          </td>
+	          <td class="align-middle">
+	          </td>                      
+	          <td class="align-middle">
+		      		<button class="buttonMoDel" type="button">수정</button>
+	          		<button class="buttonMoDel" type="button">삭제</button>
+		      </td>
+		      <td class="align-middle">
+	          </td>                     
+		    </tr>
+		    <tr>
+		      <td class="align-middle">3</td>
+		      <td class="align-middle">Larry</td>
+		      <td class="align-middle">the Bird</td>
+		      <td class="align-middle">
+		      		<button class="buttonMoDel" type="button">수정</button>
+	          		<button class="buttonMoDel" type="button">삭제</button>
+		      </td>
+		      <td class="align-middle">
+	          </td>
+	          <td class="align-middle">
+	          $69.00
+	          </td>
+	          <td class="align-middle">
+	          </td>
+	          <td class="align-middle">
+	          </td>                      
+	          <td class="align-middle">
+			      	<button class="buttonMoDel" type="button">수정</button>
+		          	<button class="buttonMoDel" type="button">삭제</button>
+		      </td>
+		      <td class="align-middle">
+	          </td>                     
+		    </tr>
+		   </tbody>
+		  </table>
+		 </div>
+    <!-- 상품 카테고리목록 끝 -->
+	</section>
+	</th:block>
+	<!-- Js Plugins -->
+	<th:block layout:fragment="customFootJs">
+	    <script src="/js/jquery-3.3.1.min.js"></script>
+	    <script src="/js/bootstrap.min.js"></script>
+	    <script src="/js/jquery.nice-select.min.js"></script>
+	    <script src="/js/jquery-ui.min.js"></script>
+	    <script src="/js/jquery.slicknav.js"></script>
+	    <script src="/js/mixitup.min.js"></script>
+	    <script src="/js/owl.carousel.min.js"></script>
+	    <script src="/js/main.js"></script>
+	</th:block>
+	<!-- Js Plugins -->
+</html>
